add skeleton of how to wire up the inference side of dialogues
Nothing actually works yet
